These changes have been submitted by ESRU and relate to the implementation of a Functional Mock-up Interface (FMI) to ESP-r.  FMI is a standard for model exchange and co-simulation and enables the development of stand-alone programs called Functional Mock-up Units (FMUs) for co-simulation with ESP-r using a master-slave protocol.
The implementation is broadly intended to allow FMUs to interact with or
replace ESP-r control functions.  At each time step during a simulation,
variables are transferred from ESP-r to the FMU, the FMU is run to make
control decisions, and control signals are transferred from the FMU back
to ESP-r to be implemented during that time step.

Detailed guides for using and developing with the FMI are provided in
ESP-rDoc/manual/FMI.

Testing:
- Tester.pl was run against the current GitHub source.  All models passed.